requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==2.0.2
certifi==2023.7.22
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.1.0
constantly==15.1.0
contourpy==1.0.7
cryptography==41.0.5
cycler==0.11.0
daphne==3.0.2
distlib==0.3.5
Django==4.1.13
django-environ==0.9.0
djangorestframework==3.13.1
filelock==3.8.0
fonttools==4.39.2
greenlet==1.1.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
msgpack==1.0.4
numpy==1.24.2
packaging==23.0
Pillow==10.0.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.3.0
pyparsing==3.1.0
PyPDF2==2.10.3
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.4
timeout-decorator==0.5.0
tomli==2.0.1
Twisted==23.10.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3<2.0
virtualenv==20.16.3
website-categorization==1.0.0
Werkzeug==2.3.8
whitenoise==6.4.0
zope.interface==5.4.0
